+++ src/app/(app)/layout.tsx
@@ -1,17 +1,13 @@
-import SideNavigation from "@/components/SideNavigation";
+import Header from "@/components/layout/Header";
 import { ReactNode } from "react";
 
-const Layout = ({ children }: { children: ReactNode }) => {
+const layout = ({ children }: { children: ReactNode }) => {
   return (
-    <>
-      <div className="flex justify-center gap-6 px-4 py-6 md:px-5 xl:py-[72px]">
-        <div className="hidden md:block">
-          <SideNavigation />
-        </div>
-        {children}
-      </div>
-    </>
+    <div>
+      <Header />
+      {children}
+    </div>
   );
 };
 
-export default Layout;
+export default layout;
